browser: simplify navigation for now
Previously, the navigation was pretty complex because of the iframe- and cursor
handling. For demonstration purposes, and because it was pretty flawed before,
this patch simplifies it by replacing with simple frontend commands.

This patch also adds TODO comments for all scrapped features, so they can be
brought back in the future.

Signed-off-by: Florian Scherf <mail@florianscherf.de>