video recorder: fix output overriding bug
Previously, when capturing to a file called `video.gif`, the video recorder
would first create a `video.mp4` and then convert it later to `video.gif`. When
`video.mp4` already existed, it got overridden.
In most cases this is not wanted behavior and is surprising to the user.

This patch changes the capturing code to just append `.mp4` to the given path
(`video.gif.mp4`) instead. This has the added benefit that this artifact is
easily to correlate with `video.gif` in case the capturing fails and milan
leaves an unclean state in the file system.

Signed-off-by: Florian Scherf <mail@florianscherf.de>