frontend: server: use actual logging instead of writing to stdout
Previously, all log, created by the frontend server, was written to stdout.
That meant that a browser could never could run completely, and the logging was
inconsistent with the rest of the project.

This patch moves all log to a sub logger of the browser instance which started
the frontend server.

Signed-off-by: Florian Scherf <mail@florianscherf.de>