video recorder: gif: remove post processing stage
Previously, gifs were created by recording to a mp4 video and converting it to
a gif in a post processing stage. This worked fine but meant handling two
processing stages and two files.

This patch introduces a new ffmpeg setup that does capturing and gif encoding
in one step. The added benefit is that the resulting gifs actually have a
higher image quality and clarity, and are smaller in file size because there is
no re encoding needed.

Signed-off-by: Florian Scherf <mail@florianscherf.de>