browser: cdp: video capturing: remove fps default unsuitable for gifs
Previously, `fps` was set to `60` by default, which was fine until

    b507668cebd9 ("video recorder: gif: make fps configurable")

because before, it was reset hard to `24` for gifs by the video recorder
internally. Now, because the fps on gifs are actually configurable, this
default produces 60 fps gifs by default, which causes slowdowns in all major
browsers.

This patch fixes this issues, by unsetting the default on the cdp browser
class, so the video recorder class can use its internal, more sane, value
of `24` fps for gifs.

Signed-off-by: Florian Scherf <mail@florianscherf.de>