utils: process: fix handling of finished processes
Previously, `Process.wait` was not called when the actual unix process exited
unexpectedly. When `Process.wait` is not called, the exit code of the unix
process never gets read, and it becomes a zombie.

To prevent zombie processes, this patch adds a `Process.wait` call right after
stdout and stderr get closed.

Signed-off-by: Florian Scherf <mail@florianscherf.de>