browser: fix error handling
Previously, the browser base class had no handling for the browser process
stopping unexpectedly in the background. When the browser stopped between two
calls, the resulting error was a broken pipe error from the websocket
connection, which is not wrong but confusing.

This patch adds handler code, that sets the browser class into an error state,
once the browser process, or the connection to the browser process, close
unexpectedly. This patch also adds a check method, which checks if the
browser is in an error state, and adds a call to this check method to every
user facing method.

Signed-off-by: Florian Scherf <mail@florianscherf.de>