tests: video capturing: fix frame rate comparison bug
Previously, the test suite sometimes crashed because the video capturing test
checked for an fps count of `24`, when the actual value, reported by ffprobe,
was a little bit lower, like `23.976023976023978` for example.

Since we do no real-time encoding, a certain margin of error is expected, and
not visible to the human eye.

This patch adds a helper function which compares two numbers with a given
margin of error in percent, and changes the test to accept an error of
one percent.

Signed-off-by: Florian Scherf <mail@florianscherf.de>