cdp client: firefox: fix handling of runtime context execution ids
Firefox requires the `contextId` parameter to be set to a runtime execution
context id when running `Runtime.evaluate`.

This patch adds code to keep track of opening and closing runtime execution
contexts, by subscribing to the events `Runtime.executionContextCreated` and
`Runtime.executionContextDestroyed`, and updates `runtime_evaluate` to use
the runtime execution context id, currently related to the top frame of the
currently attached CDP target.

Signed-off-by: Florian Scherf <mail@florianscherf.de>