browser: add browser_function decorator for better error handling
When the browser process stops or the connection to the browser process closes
down unexpectedly, the browser goes into an error state by setting
`Browser._error` to an `BrowserStoppedError`. Each public browser API function
checks whether `Browser._error` is set and if so raises the set error instead
of trying to communicate with the already closed or crashed browser.

This is an effort to have more high-level errors in these cases like
`BrowserStoppedError` instead of `websocket.ConnectionClosedError` for example.

Previously, this mechanism could run into race conditions because the browser
process could close between the check and the browser API call.

This patch adds a decorator which still implements the check if
`Browser._error` is set, but after that runs the browser API function safely
and if an exception gets raised it checks if the exception can be translated to
a more high-level exception, using `Browser.TRANSLATE_ERRORS`.

Signed-off-by: Florian Scherf <mail@florianscherf.de>