frontend: commands: frontend_function: fix function signatures
Previously, `frontend_function` was a plain decorator which overwrote function
signatures and doc strings. This is unfortunate for a public API, because it
removes documentation and makes introspection and tab-completion much harder
for developers.

This patch reimplements the decorator using `functools.wraps`, which keeps the
decorated function signature and doc string.

Signed-off-by: Florian Scherf <mail@florianscherf.de>