utils: process: add support for additional fds
`subprocess.Popen` does not implement fd mapping, which is needed for
debugging-pipes to work, because most browsers expect their parent process to
open fd 3 and 4 for them.

This patch adds the keyword `additional_fds` to `utils.process.Process`, which
uses a temp dir, fifos, and a special bash script to implement fd mapping.

Signed-off-by: Florian Scherf <mail@florianscherf.de>