cdp: video capturing: fix unstable frame rates
Previously, the video recorder class started ffmpeg when video capturing was
started and streamed all CDP screen cast frames into ffmpegs stdin, using
image2pipe. This resulted in unstable frame rates when frames were
late due to high CPU usage, because this solution used the wall clock
time of the frames arrival not its creation.

CDP provides a creation timestamp for every screen cast frame, but with the
previous solution there was no simple way to feed the incoming frames with the
timestamps into image2pipe. Therefore, this patch updates the video recorder
class to create a temporary directory when capturing is started, and to write
every frame into this directory, with the timestamp as its unix modified time.
When the video capturing is stopped, ffmpeg is invoked to render all saved
image files into a video file, using the modification times as
frame timestamps.

Signed-off-by: Florian Scherf <mail@florianscherf.de>