cdp: websocket client: fix browser info retrieval
Previously, the browser info was retrieved by just using the first CDP frame
target from the list. That worked as long as there was only one target. So no
second frame, iframe, or browser plugin.

This patch updates the retrieval code to search for the first target with the
type `page`.

Signed-off-by: Florian Scherf <mail@florianscherf.de>