frontend: cursor: awaitText: fix element not found error
`awaitText` should check whether an element with the given selector and text
exists, and retry until both requirements are met or `timeoutMax` is reached.

Previously, `awaitText` did not retry the selector properly and crashed with
an `element not found` error when no element with the given selector was found
on the first try.

Signed-off-by: Florian Scherf <mail@florianscherf.de>