frontend: awaitElement: add returnElement argument
Previously, `awaitElement` always returned the awaited element, which meant
it got encoded into an empty object into the JSON response. This is fine unless
the tested page uses a frontend framework that adds attributes to DOM elements
which end up in the JSON dump.
This caused encoding problems, especially with React sites.

This patch fixes this issue by making the return type the
function configurable.

Signed-off-by: Florian Scherf <mail@florianscherf.de>