frontend: cursor: fix cursor movement with disabled animations
Previously, the internal state of the cursor and the actual position of the
cursor image got out of sync when the cursor animations got switch on and off
between clicks.
This was due to the cursor using two conflicting ways, either
setting `element.style.top|left` or `element.animate()`, to set the cursor
position based on the animations being switch on or not.

This patch fixes this issue by only using the animation way of setting the
position, but setting the animation duration to zero in case of switched off
animations.

Signed-off-by: Florian Scherf <mail@florianscherf.de>