frontend: cursor: awaitElements: account for reloads between reloads
Previously, `Cursor.awaitElements` saved the document of a given iframe into
a variable and used it throughout the entire runtime of the function. This
works only if the iframe never reloads, otherwise it can yield false-negative
results because after a reload the new document can contain a matching element.

This patch changes the function to retrieve the document of the iframe with
every iteration of the main loop.

Signed-off-by: Florian Scherf <mail@florianscherf.de>