rename timeout to retry_interval and timeout_max to timeout
Previously, we had the terms `timeout` which described the interval in which
the frontend would try to find a selector, and `timeout_max` which described
the total time the frontend should try to find at least one matching element.

This was very confusing. Most users (including me sometimes) thought that
`timeout` describes the total amount of time the frontend should wait for a
selector to match.

This patch renames `timeout` to `retry_interval` and `timeout_max`
to `timeout`.

Signed-off-by: Florian Scherf <mail@florianscherf.de>