utils: stream: add handling for BlockingIOError and OSErrors while reading
Previously, we had hangups in Webkit while reading screenshots from a stream.
This seems to be due to blocking io in `os.read`. I am not sure why this
happens, because `utils.Stream.get_readable_stream` sets its fd
up non-blocking to avoid this very issue.

This patch adds some basic handling for this case and adds a comment that this
should be investigated in the future.

Signed-off-by: Florian Scherf <mail@florianscherf.de>